fix player collision rays position
there was a bug when sliding along a wall,
apparently because the collision box was inline with the collision rays.
Moving them a bit to the inside fixed it,
but now there could be a situation when a player is 1 pixel into a block,
because the collision box is higher than a block.